polaris_refactored/requirements.txt
# POLARIS Framework - Complete Requirements File
# Proactive Optimization & Learning Architecture for Resilient Intelligent Systems
# Version: 2.0.0
# Python: >=3.8

# ============================================================================
# Core Dependencies
# ============================================================================

# Message Bus & Async Communication
nats-py>=2.0.0,<3.0.0

# Configuration Management
PyYAML>=6.0,<7.0
python-dotenv>=0.19.0,<1.0.0

# Data Validation & Serialization
pydantic>=2.0.0,<3.0.0
jsonschema>=4.0.0,<5.0.0

# HTTP Client for LLM APIs
aiohttp>=3.8.0,<4.0.0
requests>=2.28.0,<3.0.0

# gRPC Support (optional, for advanced features)
grpcio>=1.50.0,<2.0.0
grpcio-tools>=1.50.0,<2.0.0

# ============================================================================
# LLM Integration
# ============================================================================

# Google Generative AI (Gemini)
google-generativeai>=0.3.0,<1.0.0
google-api-core>=2.10.0,<3.0.0

# ============================================================================
# Development & Testing Dependencies
# ============================================================================

# Testing Framework
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.0.0,<5.0.0

# Code Quality & Formatting
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.0.0,<2.0.0

# ============================================================================
# Optional Dependencies for Enhanced Features
# ============================================================================

# Progress bars (used in some scripts)
tqdm>=4.60.0,<5.0.0

# Rich terminal output
rich>=13.0.0

# ============================================================================
# Notes
# ============================================================================
# 
# Installation:
#   pip install -r requirements.txt
#
# Development Installation (includes dev tools):
#   pip install -r requirements.txt
#   pip install -r requirements-dev.txt
#
# For Gemini LLM Support:
#   - Set GEMINI_API_KEY environment variable in .env file
#   - google-generativeai is included in main requirements
#
# Python Version:
#   - Minimum: Python 3.8
#   - Tested: Python 3.9, 3.10, 3.11
#   - Recommended: Python 3.10+
#
# Virtual Environment Setup:
#   python -m venv .venv
#   source .venv/bin/activate  # On Windows: .venv\Scripts\activate
#   pip install -r requirements.txt
#
